+++ assembly/io/io-device.ts
@@ -1,4 +1,4 @@
 export default interface IODevice {
-    writeIO(address: u32, value: u32): void;
-    readIO(address: u32): u32;
+    writeIO(address: u32, value: u8): void;
+    readIO(address: u32): u8;
 }
